chore: changed everything according to PR suggestions TISNEW-3946
