+++ README.md
@@ -1,7 +1,7 @@
-# TIS Microservice Template
+# TIS Trainee Forms Microservice
 
 ## About
-This is a template to be used for TIS microservices with the following
+This is a TIS self service's trainee forms microservice to be used with the following
 technology:
 
  - Java 11
@@ -36,6 +36,3 @@ This project uses [Semantic Versioning](semver.org).
 
 ## License
 This project is license under [The MIT License (MIT)](LICENSE).
-
-[TemplateApplication]: src/main/java/uk/nhs/hee/trainee/forms/TraineeFormsApplication.java
-[TemplateApplicationTest]: src/test/java/uk/nhs/hee/trainee/forms/TraineeFormsApplicationTest.java
