+++ build.gradle
@@ -9,8 +9,7 @@ plugins {
   id "org.sonarqube" version "2.8"
 }
 
-// TODO: Update group to end with 'admin' or 'trainee'.
-group = 'uk.nhs.hee.trainee.forms'
+group = 'uk.nhs.hee.tis.trainee.forms'
 version = '0.0.1'
 sourceCompatibility = '11'
 
@@ -63,9 +62,7 @@ sonarqube {
     property "sonar.host.url", "https://sonarcloud.io"
     property "sonar.login", System.getenv("SONAR_TOKEN")
     property "sonar.organization", "health-education-england"
-    // TODO: Update sonar.projectKey to real value.
     property "sonar.projectKey", "Health-Education-England_tis-trainee-forms"
-
     property "sonar.java.checkstyle.reportPaths",
       "build/reports/checkstyle/main.xml,build/reports/checkstyle/test.xml"
   }
