+++ src/test/java/uk/nhs/hee/tis/trainee/forms/TraineeFormsApplicationTest.java
@@ -19,13 +19,12 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package uk.nhs.hee.trainee.forms;
+package uk.nhs.hee.tis.trainee.forms;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-//TODO: Update package and class name.
 class TraineeFormsApplicationTest {
 
   @Test
